refactor: Shared BorderLayerRenderer initialization for supported controls
