+++ src/Uno.UI/UI/Xaml/Controls/Border/Border.cs
@@ -39,6 +39,12 @@ namespace Windows.UI.Xaml.Controls
 	[ContentProperty(Name = nameof(Child))]
 	public partial class Border : FrameworkElement
 	{
+		private readonly BorderLayerRenderer _borderRenderer;
+
+		public Border()
+        {
+			_borderRenderer = new BorderLayerRenderer(this);
+		}
 
 		/// <summary>        
 		/// Support for the C# collection initializer style.
