+++ src/Uno.UI/UI/Xaml/Controls/Border/Border.iOSmacOS.cs
@@ -20,15 +20,6 @@ namespace Windows.UI.Xaml.Controls
 {
 	public partial class Border
 	{
-		private readonly BorderLayerRenderer _borderRenderer;
-
-		public Border()
-		{
-			_borderRenderer = new BorderLayerRenderer(this);
-
-			this.RegisterLoadActions(() => UpdateBorderLayer(), () => _borderRenderer.Clear());
-		}
-
         partial void OnBorderBrushChangedPartial()
 		{
 			UpdateBorderLayer();
