+++ src/Uno.UI/UI/Xaml/Controls/ContentPresenter/ContentPresenter.Android.cs
@@ -20,14 +20,8 @@ namespace Windows.UI.Xaml.Controls
 {
 	public partial class ContentPresenter
 	{
-		private readonly BorderLayerRenderer _borderRenderer;
-
-		public ContentPresenter()
+		partial void InitializePlatform()
 		{
-			_borderRenderer = new BorderLayerRenderer(this);
-
-			InitializeContentPresenter();
-
 			IFrameworkElementHelper.Initialize(this);
 		}
