+++ src/Uno.UI/UI/Xaml/Controls/ContentPresenter/ContentPresenter.iOS.cs
@@ -18,15 +18,6 @@ namespace Windows.UI.Xaml.Controls
 	/// </remarks>
 	public partial class ContentPresenter
 	{
-		private readonly BorderLayerRenderer _borderRenderer;
-
-		public ContentPresenter()
-		{
-			_borderRenderer = new BorderLayerRenderer(this);
-
-			InitializeContentPresenter();
-		}
-
 		private void SetUpdateTemplate()
 		{
 			UpdateContentTemplateRoot();
