+++ src/Uno.UI/UI/Xaml/Controls/ContentPresenter/ContentPresenter.macOS.cs
@@ -17,15 +17,6 @@ namespace Windows.UI.Xaml.Controls
 	/// </remarks>
 	public partial class ContentPresenter
 	{
-		private readonly BorderLayerRenderer _borderRenderer;
-
-		public ContentPresenter()
-		{
-			_borderRenderer = new BorderLayerRenderer(this);
-
-			InitializeContentPresenter();
-		}
-
 		private void SetUpdateTemplate()
 		{
 			UpdateContentTemplateRoot();
