+++ src/Uno.UI/UI/Xaml/Controls/ContentPresenter/ContentPresenter.net.cs
@@ -17,11 +17,6 @@ namespace Windows.UI.Xaml.Controls
 	/// </remarks>
 	public partial class ContentPresenter : FrameworkElement
 	{
-		public ContentPresenter()
-		{
-			InitializeContentPresenter();
-		}
-
 		private void SetUpdateTemplate()
 		{
 			UpdateContentTemplateRoot();
