+++ src/Uno.UI/UI/Xaml/Controls/ContentPresenter/ContentPresenter.skia.cs
@@ -19,21 +19,13 @@ namespace Windows.UI.Xaml.Controls
 	/// </remarks>
 	public partial class ContentPresenter : FrameworkElement
 	{
-		private readonly BorderLayerRenderer _borderRenderer;
-		
 		private Rect? _lastArrangeRect;
 		private Rect _lastGlobalRect;
 		private bool _nativeHostRegistered;
 
 		public ContentPresenter()
 		{
-			_borderRenderer = new BorderLayerRenderer(this);
-
 			InitializeContentPresenter();
-
-			Loaded += (s, e) => RegisterNativeHostSupport();
-			Unloaded += (s, e) => UnregisterNativeHostSupport();
-			LayoutUpdated += (s, e) => UpdateBorder();
 		}
 
 		private void SetUpdateTemplate()
