+++ src/Uno.UI/UI/Xaml/Controls/Page/Page.cs
@@ -7,9 +7,11 @@ namespace Windows.UI.Xaml.Controls
 {
 	public partial class Page : UserControl
 	{
+		private readonly BorderLayerRenderer _borderRenderer;
+
 		public Page()
 		{
-			InitializeBorder();
+			_borderRenderer = new BorderLayerRenderer(this);
 		}
 
 		protected internal virtual void OnNavigatedFrom(NavigationEventArgs e) { }
