+++ src/Uno.UI/UI/Xaml/Controls/Page/Page.iOS.cs
@@ -11,17 +11,6 @@ namespace Windows.UI.Xaml.Controls
 {
 	public partial class Page
 	{
-		private BorderLayerRenderer _borderRenderer;
-
-		private void InitializeBorder()
-		{
-			_borderRenderer = new BorderLayerRenderer(this);
-
-			Loaded += (s, e) => UpdateBorder();
-			Unloaded += (s, e) => _borderRenderer.Clear();
-			LayoutUpdated += (s, e) => UpdateBorder();
-		}
-
 		public override void LayoutSubviews()
 		{
 			base.LayoutSubviews();
