+++ src/Uno.UI/UI/Xaml/Controls/Page/Page.macOS.cs
@@ -10,17 +10,6 @@ namespace Windows.UI.Xaml.Controls
 {
 	public partial class Page
 	{
-		private BorderLayerRenderer _borderRenderer;
-
-		private void InitializeBorder()
-		{
-			_borderRenderer = new BorderLayerRenderer(this);
-
-			Loaded += (s, e) => UpdateBorder();
-			Unloaded += (s, e) => _borderRenderer.Clear();
-			LayoutUpdated += (s, e) => UpdateBorder();
-		}
-
 		public override void Layout()
 		{
 			base.Layout();
