+++ src/Uno.UI/UI/Xaml/Controls/Page/Page.net.cs
@@ -10,10 +10,6 @@ namespace Windows.UI.Xaml.Controls
 {
 	public partial class Page
 	{
-		private void InitializeBorder()
-		{
-		}
-
 		private void UpdateBorder()
 		{
 		}
