+++ src/Uno.UI/UI/Xaml/Controls/Page/Page.skia.cs
@@ -10,17 +10,6 @@ namespace Windows.UI.Xaml.Controls
 {
 	public partial class Page
 	{
-		private BorderLayerRenderer _borderRenderer;
-
-		private void InitializeBorder()
-		{
-			_borderRenderer = new BorderLayerRenderer(this);
-
-			Loaded += (s, e) => UpdateBorder();
-			Unloaded += (s, e) => _borderRenderer.Clear();
-			LayoutUpdated += (s, e) => UpdateBorder();
-		}
-
 		private void UpdateBorder()
 		{
 			if (IsLoaded)
