+++ src/Uno.UI/UI/Xaml/Controls/Page/Page.wasm.cs
@@ -10,16 +10,6 @@ namespace Windows.UI.Xaml.Controls
 {
 	public partial class Page
 	{
-		private BorderLayerRenderer _borderRenderer;
-
-		private void InitializeBorder()
-		{
-			_borderRenderer ??= new BorderLayerRenderer(this);
-
-			Loaded += (s, e) => UpdateBorder();
-			Unloaded += (s, e) => _borderRenderer.Clear();
-		}
-
 		private void UpdateBorder()
 		{
 			if (IsLoaded)
