+++ src/Uno.UI/UI/Xaml/Controls/Panel/Panel.skia.cs
@@ -21,22 +21,6 @@ namespace Windows.UI.Xaml.Controls
 {
 	public partial class Panel : IEnumerable
 	{
-		private SerialDisposable _brushChanged = new SerialDisposable();
-		private readonly BorderLayerRenderer _borderRenderer;
-
-		public Panel()
-		{
-			_borderRenderer = new BorderLayerRenderer(this);
-
-			Initialize();
-
-			Loaded += (s, e) => UpdateBorder();
-			Unloaded += (s, e) => _borderRenderer.Clear();
-			LayoutUpdated += (s, e) => UpdateBorder();
-		}
-
-		partial void Initialize();
-
 		partial void UpdateBorder()
 		{
 			// Checking for Window avoids re-creating the layer until it is actually used.
